remove label selector from DNSServiceConfig
The label selector is incompatible with the current implementation that loads the config
on-demand during each reconciliation.
We would either need to load it once during startup, which would mean that the controller
needs to be restarted for changes to take effect,
or we would need a more complicated setup with an own controller for the config.

Since clusters can already be filtered by their purposes, there is probably no need for
a label selector anyway, at least not for now.