feat: external-dns cluster controller (#12)
* deploy HelmRelease for external-dns

* finish cluster controller implementation

* actually start cluster controller

* reconcile clusters if external-dns config changes (requires controller-utils update)

* remove label selector from DNSServiceConfig

The label selector is incompatible with the current implementation that loads the config
on-demand during each reconciliation.
We would either need to load it once during startup, which would mean that the controller
needs to be restarted for changes to take effect,
or we would need a more complicated setup with an own controller for the config.

Since clusters can already be filtered by their purposes, there is probably no need for
a label selector anyway, at least not for now.

* fix bugs and add first unit tests

* add more unit tests

* remove library function which should not be used in this context

* fix linting issues

* fix helm release

* fix controller commands

* fix DNSServiceConfig being cluster-scoped

* enable copying any amount of secrets

* implement helm values replace feature

* make purposeSelector behavior more consistent

* add documentation

* bump build submodule

* implement review feedback